added avltree ,double modified,some fixes

+++ project/APMS/classes/area.hpp
@@ -35,8 +35,8 @@ public:
     void Print_Area_By_Year(int year,int choice);
     void Print_Area_By_month(int year,int month);
     //resumed print 
-    int get_area_total_sales_per_year(int year);
-    int get_area_total_sales_per_month(int year,int month);
+    double get_area_total_sales_per_year(int year);
+    double get_area_total_sales_per_month(int year,int month);
 
 
     friend ostream& operator<<(ostream&, const Area&);
@@ -49,11 +49,11 @@ public:
     void print_Area_monthly_penalty(int year,int month,string category_name);
     void print_Area_yearly_penalty(int year,string category_name);
     //functions to get the water and electricity consumption of the area
-    int get_area_yearly_water_consumption(int year);
-    int get_area_yearly_electricity_consumption(int year);
+    double get_area_yearly_water_consumption(int year);
+    double get_area_yearly_electricity_consumption(int year);
     //functions to get the water and electricity consumption of the area in a month
-    int get_area_monthly_water_consumption(int year,int month);
-    int get_area_monthly_electricity_consumption(int year,int month);
+    double get_area_monthly_water_consumption(int year,int month);
+    double get_area_monthly_electricity_consumption(int year,int month);
     
 
 };
